Implement functionality to add token to board
This adds a 'Board' class that, when initialized,
represents a 'Connect Four' game grid with a
width of 7 and height of 6. The #insert method
takes a column index and a token and attempts to
insert the token to the bottom-most spot of the
column, only if the column index is in range and
the column is not yet full. Unit tests were also
included for the #initialize and #insert methods.