Fix bug for vertical win condition on board
Bug existed where as long as there were four
tokens in a column, it would be counted as a win,
even if they were not connected. This commit fixes
the bug and adds unit tests testing the behavior
when four tokens are connected and not connected
in a given column.